+++ main/java/servlet/admin/RemoveIntegratedBankServlet.java
@@ -3,6 +3,7 @@ package servlet.admin;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
+import java.util.Collection;
 import java.util.LinkedList;
 
 import javax.servlet.ServletException;
@@ -14,51 +15,59 @@ import javax.servlet.http.HttpServletResponse;
 import dao.IntegratedBankDAO;
 import model.IntegratedBank;
 import util.Factory;
+import util.Util;
 
 public class RemoveIntegratedBankServlet extends HttpServlet {
-	private static final long serialVersionUID = -4181185294048918446L;
-	private IntegratedBankDAO integratedBankDAO;
-	
-	public void init() {
-		integratedBankDAO = Factory.getIntegratedBankDAO();
-	}
-	
-	
 	public void doGet(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
-		LinkedList<IntegratedBank> integratedBanks;
-		
-		try {
-			integratedBanks = integratedBankDAO.getAll();
-			req.setAttribute("values", integratedBanks);
-			req.getRequestDispatcher("/jsp/admin/removeIntegratedBank.jsp").include(req, res);
-		} catch(SQLException e) {
-			res.setStatus(500);
-			res.getWriter().println("<h1>Internal server error</h1>");
-		}
+		IntegratedBankDAO integratedBankDAO = Factory.getIntegratedBankDAO();
+		Collection<IntegratedBank> integratedBanks = integratedBankDAO.getAll();
+		req.setAttribute("values", integratedBanks);
+		req.getRequestDispatcher("/jsp/admin/removeIntegratedBank.jsp").include(req, res);
 	}
 	
 	
 	public void doPost(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
+		IntegratedBankDAO integratedBankDAO = Factory.getIntegratedBankDAO();
 		PrintWriter out = res.getWriter();
+		
+		IntegratedBank bank = null;
+		boolean isError = false, exceptionOccured = false;
+		String msg = "";
 		int integratedBankId;
-		boolean status = false;
 		
 		try {
 			integratedBankId = Integer.parseInt(req.getParameter("bank-id"));
-			status = integratedBankDAO.delete(integratedBankId);
 			
-			if(status)
-				out.println("<div class='notification success'>" + "integrated bank deleted successfully" + "</div>");
-			else
-				out.println("<div class='notification danger'>" + "invalid bank selected" + "</div>");
-		
+			bank = integratedBankDAO.get(integratedBankId);
+			
+			if(bank == null) {
+				isError = true;
+				msg = "Integrated bank does not exist !!!";
+			}
+			
+			if(!isError) {
+				integratedBankDAO.delete(integratedBankId);
+				msg = "integrated bank deleted successfully";				
+			}
+			
 		} catch(NumberFormatException e) {
-			out.println("<div class='notification danger'>" + "invalid input" + "</div>");
+			System.out.println(e.getMessage());
+			exceptionOccured = true;
+			msg = "Invalid input !!!";
 		} catch(SQLException e) {
-			out.println("<div class='notification danger'>" + e.getMessage() + "</div>");
+			System.out.println(e.getMessage());
+			exceptionOccured = true;
+			msg = e.getMessage();
 		} finally {
-			doGet(req, res);
-			out.close();
+			
+			if(isError || exceptionOccured) {
+				out.println(Util.createNotification(msg, "danger"));
+				doGet(req, res);
+				out.close();
+			} else {
+				res.sendRedirect(String.format("/bank-app/admin/integrated-banks?msg=%s&status=success", msg));
+			}
+			
 		}
 	}
 }
